apps/crawler/firecrowl/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "website-crawler"
version = "1.0.0"
description = "Modern web scraping with OCR support using Firecrawl API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Organization", email = "your-email@example.com"}
]
keywords = ["web-scraping", "ocr", "firecrawl", "crawler"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "firecrawl-py>=1.0.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.9.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "paddleocr>=2.6.0",
    "boto3>=1.20.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "isort>=5.0.0",
    "mypy>=0.910",
    "flake8>=3.9.0",
]

[project.scripts]
website-crawler = "website_crawler.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/firecrawl-website-crawler"
Documentation = "https://github.com/your-org/firecrawl-website-crawler#readme"
Repository = "https://github.com/your-org/firecrawl-website-crawler.git"
Issues = "https://github.com/your-org/firecrawl-website-crawler/issues"

[tool.setuptools.packages.find]
include = ["website_crawler*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["website_crawler"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true